Enum libp2p_swarm::behaviour::FromSwarm

pub enum FromSwarm<'a, Handler: IntoConnectionHandler> {
    ConnectionEstablished(ConnectionEstablished<'a>),
    ConnectionClosed(ConnectionClosed<'a, Handler>),
    AddressChange(AddressChange<'a>),
    DialFailure(DialFailure<'a, Handler>),
    ListenFailure(ListenFailure<'a, Handler>),
    NewListener(NewListener),
    NewListenAddr(NewListenAddr<'a>),
    ExpiredListenAddr(ExpiredListenAddr<'a>),
    ListenerError(ListenerError<'a>),
    ListenerClosed(ListenerClosed<'a>),
    NewExternalAddr(NewExternalAddr<'a>),
    ExpiredExternalAddr(ExpiredExternalAddr<'a>),
}

Enumeration with the list of the possible events to pass to on_swarm_event.

Variants

ConnectionEstablished(ConnectionEstablished<'a>)

Informs the behaviour about a newly established connection to a peer.

ConnectionClosed(ConnectionClosed<'a, Handler>)

Informs the behaviour about a closed connection to a peer.

This event is always paired with an earlier FromSwarm::ConnectionEstablished with the same peer ID, connection ID and endpoint.

AddressChange(AddressChange<'a>)

Informs the behaviour that the ConnectedPoint of an existing connection has changed.

DialFailure(DialFailure<'a, Handler>)

Informs the behaviour that the dial to a known or unknown node failed.

ListenFailure(ListenFailure<'a, Handler>)

Informs the behaviour that an error happened on an incoming connection during its initial handshake.

This can include, for example, an error during the handshake of the encryption layer, or the connection unexpectedly closed.

NewListener(NewListener)

Informs the behaviour that a new listener was created.

NewListenAddr(NewListenAddr<'a>)

Informs the behaviour that we have started listening on a new multiaddr.

ExpiredListenAddr(ExpiredListenAddr<'a>)

Informs the behaviour that a multiaddr we were listening on has expired, which means that we are no longer listening on it.

ListenerError(ListenerError<'a>)

Informs the behaviour that a listener experienced an error.

ListenerClosed(ListenerClosed<'a>)

Informs the behaviour that a listener closed.

NewExternalAddr(NewExternalAddr<'a>)

Informs the behaviour that we have discovered a new external address for us.

ExpiredExternalAddr(ExpiredExternalAddr<'a>)

Informs the behaviour that an external address was removed.