Struct libp2p_swarm::ExpandedSwarm

pub struct ExpandedSwarm<TTransport, TBehaviour, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo = PeerId> where
    TTransport: Transport,  { /* fields omitted */ }

Contains the state of the network, plus the way it should behave.

Methods

impl<TTransport, TBehaviour, TMuxer, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo> ExpandedSwarm<TTransport, TBehaviour, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo> where
TBehaviour: NetworkBehaviour<ProtocolsHandler = THandler>,
TInEvent: Send + 'static,
TOutEvent: Send + 'static,
TConnInfo: ConnectionInfo<PeerId = PeerId> + Debug + Clone + Send + 'static,
TMuxer: StreamMuxer + Send + Sync + 'static,
TMuxer::OutboundSubstream: Send + 'static,
TMuxer::Substream: Send + 'static,
TTransport: Transport<Output = (TConnInfo, TMuxer)> + Clone,
TTransport::Error: Send + 'static,
TTransport::Listener: Send + 'static,
TTransport::ListenerUpgrade: Send + 'static,
TTransport::Dial: Send + 'static,
THandlerErr: Error + Send + 'static,
THandler: IntoProtocolsHandler + Send + 'static,
THandler::Handler: ProtocolsHandler<InEvent = TInEvent, OutEvent = TOutEvent, Substream = Substream<TMuxer>, Error = THandlerErr> + Send + 'static,
<THandler::Handler as ProtocolsHandler>::OutboundOpenInfo: Send + 'static,
<THandler::Handler as ProtocolsHandler>::InboundProtocol: InboundUpgrade<Negotiated<Substream<TMuxer>>> + Send + 'static,
<<THandler::Handler as ProtocolsHandler>::InboundProtocol as UpgradeInfo>::Info: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::InboundProtocol as UpgradeInfo>::InfoIter: Send + 'static,
<<<THandler::Handler as ProtocolsHandler>::InboundProtocol as UpgradeInfo>::InfoIter as IntoIterator>::IntoIter: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::InboundProtocol as InboundUpgrade<Negotiated<Substream<TMuxer>>>>::Error: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::InboundProtocol as InboundUpgrade<Negotiated<Substream<TMuxer>>>>::Future: Send + 'static,
<THandler::Handler as ProtocolsHandler>::OutboundProtocol: OutboundUpgrade<Negotiated<Substream<TMuxer>>> + Send + 'static,
<<THandler::Handler as ProtocolsHandler>::OutboundProtocol as UpgradeInfo>::Info: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::OutboundProtocol as UpgradeInfo>::InfoIter: Send + 'static,
<<<THandler::Handler as ProtocolsHandler>::OutboundProtocol as UpgradeInfo>::InfoIter as IntoIterator>::IntoIter: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::OutboundProtocol as OutboundUpgrade<Negotiated<Substream<TMuxer>>>>::Future: Send + 'static,
<<THandler::Handler as ProtocolsHandler>::OutboundProtocol as OutboundUpgrade<Negotiated<Substream<TMuxer>>>>::Error: Send + 'static,

pub fn new(
transport: TTransport,
behaviour: TBehaviour,
local_peer_id: PeerId
) -> Self

Builds a new Swarm.

pub fn transport(me: &Self) -> &TTransport

Returns the transport passed when building this object.

pub fn listen_on(
me: &mut Self,
addr: Multiaddr
) -> Result<ListenerId, TransportError<TTransport::Error>>

Starts listening on the given address.

Returns an error if the address is not supported.

pub fn remove_listener(me: &mut Self, id: ListenerId) -> Result<(), ()>

Remove some listener.

Returns Ok(()) if there was a listener with this ID.

pub fn dial_addr(
me: &mut Self,
addr: Multiaddr
) -> Result<(), TransportError<TTransport::Error>>

Tries to dial the given address.

Returns an error if the address is not supported.

pub fn dial(me: &mut Self, peer_id: PeerId)

Tries to reach the given peer using the elements in the topology.

Has no effect if we are already connected to that peer, or if no address is known for the peer.

pub fn listeners(me: &Self) -> impl Iterator<Item = &Multiaddr>

Returns an iterator that produces the list of addresses we're listening on.

pub fn external_addresses(me: &Self) -> impl Iterator<Item = &Multiaddr>

Returns an iterator that produces the list of addresses that other nodes can use to reach us.

pub fn local_peer_id(me: &Self) -> &PeerId

Returns the peer ID of the swarm passed as parameter.

pub fn add_external_address(me: &mut Self, addr: Multiaddr)

Adds an external address.

An external address is an address we are listening on but that accounts for things such as NAT traversal.

pub fn connection_info(me: &mut Self, peer_id: &PeerId) -> Option<TConnInfo>

Returns the connection info of a node, or None if we're not connected to it.

pub fn ban_peer_id(me: &mut Self, peer_id: PeerId)

Bans a peer by its peer ID.

Any incoming connection and any dialing attempt will immediately be rejected. This function has no effect is the peer is already banned.

pub fn unban_peer_id(me: &mut Self, peer_id: PeerId)

Unbans a peer.

pub async fn next_event<'_>(&'_ mut self) -> SwarmEvent<TBehaviour::OutEvent>

Returns the next event that happens in the Swarm.

Includes events from the NetworkBehaviour but also events about the connections status.

pub async fn next<'_>(&'_ mut self) -> TBehaviour::OutEvent

Returns the next event produced by the NetworkBehaviour.