Struct libp2p_swarm::OneShotHandler

pub struct OneShotHandler<TInbound, TOutbound, TEvent>where
    TOutbound: OutboundUpgradeSend,{ /* private fields */ }

A ConnectionHandler that opens a new substream for each request.

Implementations

impl<TInbound, TOutbound, TEvent> OneShotHandler<TInbound, TOutbound, TEvent>where
TOutbound: OutboundUpgradeSend,

pub fn new(
listen_protocol: SubstreamProtocol<TInbound, ()>,
config: OneShotHandlerConfig
) -> Self

Creates a OneShotHandler.

pub fn pending_requests(&self) -> u32

Returns the number of pending requests.

pub fn listen_protocol_ref(&self) -> &SubstreamProtocol<TInbound, ()>

Returns a reference to the listen protocol configuration.

Note: If you modify the protocol, modifications will only applies to future inbound substreams, not the ones already being negotiated.

pub fn listen_protocol_mut(&mut self) -> &mut SubstreamProtocol<TInbound, ()>

Returns a mutable reference to the listen protocol configuration.

Note: If you modify the protocol, modifications will only applies to future inbound substreams, not the ones already being negotiated.

pub fn send_request(&mut self, upgrade: TOutbound)

Opens an outbound substream with upgrade.