Struct libp2p_swarm::Swarm

pub struct Swarm<TBehaviour>where
    TBehaviour: NetworkBehaviour,{ /* private fields */ }

Contains the state of the network, plus the way it should behave.

Note: Needs to be polled via <Swarm as Stream> in order to make progress.

Implementations

impl<TBehaviour> Swarm<TBehaviour>where
TBehaviour: NetworkBehaviour,

pub fn new(
transport: Boxed<(PeerId, StreamMuxerBox)>,
behaviour: TBehaviour,
local_peer_id: PeerId
) -> Self

Builds a new Swarm.

pub fn network_info(&self) -> NetworkInfo

Returns information about the connections underlying the Swarm.

pub fn listen_on(
&mut self,
addr: Multiaddr
) -> Result<ListenerId, TransportError<Error>>

Starts listening on the given address. Returns an error if the address is not supported.

Listeners report their new listening addresses as SwarmEvent::NewListenAddr. Depending on the underlying transport, one listener may have multiple listening addresses.

pub fn remove_listener(&mut self, listener_id: ListenerId) -> bool

Remove some listener.

Returns true if there was a listener with this ID, false otherwise.

pub fn dial(&mut self, opts: impl Into<DialOpts>) -> Result<(), DialError>

Dial a known or unknown peer.

See also DialOpts.

let mut swarm = Swarm::new(
  DummyTransport::new().boxed(),
  DummyBehaviour::default(),
  PeerId::random(),
);

// Dial a known peer.
swarm.dial(PeerId::random());

// Dial an unknown peer.
swarm.dial("/ip6/::1/tcp/12345".parse::<Multiaddr>().unwrap());

pub fn listeners(&self) -> impl Iterator<Item = &Multiaddr>

Returns an iterator that produces the list of addresses we’re listening on.

pub fn local_peer_id(&self) -> &PeerId

Returns the peer ID of the swarm passed as parameter.

pub fn external_addresses(&self) -> impl Iterator<Item = &AddressRecord>

Returns an iterator for AddressRecords of external addresses of the local node, in decreasing order of their current score.

pub fn add_external_address(
&mut self,
a: Multiaddr,
s: AddressScore
) -> AddAddressResult

Adds an external address record for the local node.

An external address is an address of the local node known to be (likely) reachable for other nodes, possibly taking into account NAT. The external addresses of the local node may be shared with other nodes by the NetworkBehaviour.

The associated score determines both the position of the address in the list of external addresses (which can determine the order in which addresses are used to connect to) as well as how long the address is retained in the list, depending on how frequently it is reported by the NetworkBehaviour via NetworkBehaviourAction::ReportObservedAddr or explicitly through this method.

pub fn remove_external_address(&mut self, addr: &Multiaddr) -> bool

Removes an external address of the local node, regardless of its current score. See Swarm::add_external_address for details.

Returns true if the address existed and was removed, false otherwise.

pub fn ban_peer_id(&mut self, peer_id: PeerId)

Bans a peer by its peer ID.

Any incoming connection and any dialing attempt will immediately be rejected. This function has no effect if the peer is already banned.

pub fn unban_peer_id(&mut self, peer_id: PeerId)

Unbans a peer.

pub fn disconnect_peer_id(&mut self, peer_id: PeerId) -> Result<(), ()>

Disconnects a peer by its peer ID, closing all connections to said peer.

Returns Ok(()) if there was one or more established connections to the peer.

Note: Closing a connection via Swarm::disconnect_peer_id does not inform the corresponding ConnectionHandler. Closing a connection via a ConnectionHandler can be done either in a collaborative manner across ConnectionHandlers with ConnectionHandler::connection_keep_alive or directly with ConnectionHandlerEvent::Close.

pub fn is_connected(&self, peer_id: &PeerId) -> bool

Checks whether there is an established connection to a peer.

pub fn connected_peers(&self) -> impl Iterator<Item = &PeerId>

Returns the currently connected peers.

pub fn behaviour(&self) -> &TBehaviour

Returns a reference to the provided NetworkBehaviour.

pub fn behaviour_mut(&mut self) -> &mut TBehaviour

Returns a mutable reference to the provided NetworkBehaviour.