Type Definition libp2p_swarm::PendingOutboundConnectionError

pub type PendingOutboundConnectionError<TTransErr> = PendingConnectionError<Vec<(Multiaddr, TransportError<TTransErr>)>>;

Errors that can occur in the context of a pending outgoing Connection.

Note: Addresses for an outbound connection are dialed in parallel. Thus, compared to PendingInboundConnectionError, one or more TransportErrors can occur for a single connection.