Crate libp2p_swarm

High level manager of the network.

A Swarm contains the state of the network as a whole. The entire behaviour of a libp2p network can be controlled through the Swarm. The Swarm struct contains all active and pending connections to remotes and manages the state of all the substreams that have been opened, and all the upgrades that were built upon these substreams.

Initializing a Swarm

Creating a Swarm requires three things:

A network identity of the local node in form of a PeerId.

An implementation of the Transport trait. This is the type that will be used in order to reach nodes on the network based on their address. See the transport module for more information.

An implementation of the NetworkBehaviour trait. This is a state machine that defines how the swarm should behave once it is connected to a node.

Network Behaviour

The NetworkBehaviour trait is implemented on types that indicate to the swarm how it should behave. This includes which protocols are supported and which nodes to try to connect to. It is the NetworkBehaviour that controls what happens on the network. Multiple types that implement NetworkBehaviour can be composed into a single behaviour.

Protocols Handler

The ConnectionHandler trait defines how each active connection to a remote should behave: how to handle incoming substreams, which protocols are supported, when to open a new outbound substream, etc.

Re-exports

pub use behaviour::CloseConnection;

pub use behaviour::NetworkBehaviour;

pub use behaviour::NetworkBehaviourAction;

pub use behaviour::NetworkBehaviourEventProcess;

pub use behaviour::NotifyHandler;

pub use behaviour::PollParameters;

pub use handler::ConnectionHandler;

pub use handler::ConnectionHandlerEvent;

pub use handler::ConnectionHandlerUpgrErr;

pub use handler::IntoConnectionHandler;

pub use handler::KeepAlive;

pub use handler::SubstreamProtocol;

Modules

behaviour

dial_opts

handler

Once a connection to a remote peer is established, a ConnectionHandler negotiates and handles one or more specific protocols on the connection.

Structs

AddressRecord

An record in a prioritised list of addresses.

ConnectionCounters

Network connection information.

ConnectionHandlerSelect

Implementation of ConnectionHandler that combines two protocols into one.

ConnectionLimit

Information about a connection limit.

ConnectionLimits

The configurable connection limits.

DummyBehaviour

Dummy implementation of NetworkBehaviour that doesn’t do anything.

IntoConnectionHandlerSelect

Implementation of IntoConnectionHandler that combines two protocols into one.

NetworkInfo

Information about the connections obtained by Swarm::network_info().

OneShotHandler

A ConnectionHandler that opens a new substream for each request.

OneShotHandlerConfig

Configuration parameters for the OneShotHandler

Swarm

Contains the state of the network, plus the way it should behave.

SwarmBuilder

A SwarmBuilder provides an API for configuring and constructing a Swarm.

SwarmPollParameters

Parameters passed to poll(), that the NetworkBehaviour has access to.

Enums

AddAddressResult

The result of adding an address to an ordered list of addresses with associated scores.

AddressScore

The “score” of an address w.r.t. an ordered collection of addresses.

ConnectionError

Errors that can occur in the context of an established Connection.

DialError

The possible failures of dialing.

PendingConnectionError

Errors that can occur in the context of a pending Connection.

SwarmEvent

Event generated by the Swarm.

Type Definitions

NegotiatedSubstream

Substream for which a protocol has been chosen.

PendingInboundConnectionError

Errors that can occur in the context of a pending incoming Connection.

PendingOutboundConnectionError

Errors that can occur in the context of a pending outgoing Connection.